pyproject.toml
[project]
name = "synthe"
version = "0.1.0"
description = "SYNTHE: Synthesis of Online Learning Primitives for Hierarchical Temporal Engines — A mechanistic alternative to Transformers"
authors = [{ name = "Kael Valen" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "deep-learning",
    "sequence-modeling",
    "state-space-models",
    "kalman-filter",
    "associative-memory",
    "online-learning",
    "language-model",
    "pytorch",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1.0",
    "einops>=0.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-benchmark",
]
train = [
    "datasets>=2.14.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
]
eval = [
    "lm-eval>=0.4.0",
    "datasets>=2.14.0",
]

[project.urls]
Homepage = "https://github.com/kaelvalen/synthe"
Repository = "https://github.com/kaelvalen/synthe"
Issues = "https://github.com/kaelvalen/synthe/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
